refactor!: Patch Sidekiq::BasicFetch
Do not provide custom BasicFetch. Patch upstream instead.
This should make integration with Sidkiq-Pro easier.
